etl_be/requirements.txt
pandas>=2.2.0
pika>=1.3.2
python-dotenv>=1.0.1
psycopg2-binary>=2.9.0
pytest>=8.2.0
sqlalchemy>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

